% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregiereFirmen.R
\name{aggregiereFirmen}
\alias{aggregiereFirmen}
\title{aggregiereFirmen}
\usage{
aggregiereFirmen(Firmenanalyse, alpha)
}
\arguments{
\item{Firmenanalyse}{Eine Liste, die mit analysiereFirmen() erzeugt wurde.}
}
\value{
Eine Liste mit Statistiken und Teststatistiken. AR_quer_t, varHat_AR_quer_t, CAR_quer, varHat_Car_quuer_t sowie zwei Werte für die Konfidenzbänder
}
\description{
Aggregiert die Ergebnisse auf Firmenebene
}
\details{
Aggregiert die Ergebnisse, die per analysiereFirmen() generiert werdne
}
\author{
Harald Fiedler
}

