% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysiereFirmen.R
\name{analysiereFirmen}
\alias{analysiereFirmen}
\title{analysiereFirmen}
\usage{
analysiereFirmen(Gruppe, eventdate, Puffer = 50, left = -5, right = 5,
  L_est = 200, alpha = 0.01, inProzent = TRUE)
}
\arguments{
\item{Gruppe}{character-Vektor der Länge 1 aus c("A", "B", "C", "AandB", "AxorB"). Gibt an, welche Firmengruppe untersucht werden soll. "A" und "B" haben unterschiedlichen Umgang mit Pensionsrisiken. "C" hat keine Pensionen und ist eine Art Kontrollgruppe. Da manche Firmen über mehrere Jahe teilweise Strategie zwischen "A" und "B" gewechselt haben, ist "AandB" und "AxorB" evtl. acuh von Interesse.}

\item{eventdate}{eine Zahl aus c(260, 555, 646, 809, 958), die angibt, am wievielten Handelstag ein besonderes Event stattgefunden hat.}

\item{Puffer}{numeric der Länge 1, üblicherweise aus c(0,50). Führ ggf. einen Space zwischen dem estimation Window und dem event Window ein. Defualt ist 50.}

\item{left}{gibt an, wie viele Tage vor dem Event das Event-Window beginnen soll.}

\item{right}{gibt an, wie viele Tage nach dem Event das Event-Window beginnen soll. Übliche Windows sind [-5, 5], [-1,1], aber auch [-3, 3]}

\item{L_est}{eine Zahl, die angibt, wie lange das Estimation-Window sein soll. Default ist 200.}

\item{alpha}{Das Konfidenzniveau, default ist 1%}

\item{inProzent}{Boolean der Länge 1, das angibt, ob die Aktienkurse in prozentualen Veränderungen (TRUE) oder in absoluten Beträgen untersucht werden sollen.}
}
\value{
Eine Liste von zwei Listen. Die Liste AR_it enthält die AR_it, die Konfidenzgrenzen oben, die Konfidenzgrenzen unten und die Schätzer für die Residuenvarianz. Die zweite Liste das gleiche für die Kummulanten.
}
\description{
Diese Funktion analysiert die Daten und liefert Ergebnisse auf Ebene der individuellen Firmen.
}
\details{
Diese Funktion ist das Arbeitspferd des Pakets. Sie nimmt die Daten und berechnet ein Marktmodell, und gibt darauf basierend die abnormal Returns und andere Kennwerte aus. Die Analyse beschränkt sich dabei vollständig auf individuelle Firmen, führt also keinerlei Aggregierung der Ergebnisse durch. Für die Aggreggierung siehe analysiereMarkt().
}
\author{
Harald Fiedler
}

