% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEventFrame.R
\name{buildEventFrame}
\alias{buildEventFrame}
\title{buildEventFrame}
\usage{
buildEventFrame(eventdate, Puffer, left, right, L_est, ISIN, KURS)
}
\arguments{
\item{Es}{werden lediglich Variabeln aus dem Wrapper durchgereicht.}
}
\value{
Ein data.frame. Reduziert den Datensatz KURS auf die in ISIN indizierten Firmen. Dazu kommen noch boolesche Zugriffsschlüssel und ein zeitlicher Index.
}
\description{
Hilfsfuntkion.
}
\details{
Eine interne Hilfsfunktion von analysiereFirmen().
}
\author{
Harald Fiedler
}

