% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AR_it.R
\name{calc_AR_it}
\alias{calc_AR_it}
\title{calc_AR_it}
\usage{
calc_AR_it(Firma, estimationWindow, eventWindow, Market)
}
\arguments{
\item{Werden}{aus dem Wrapper durchgereicht.}
}
\value{
Ein data.frame() wie in buildEventFrame(), aber in den Werten um den bedingten Erwartungswert bereinigt, also die Residuen enthaltend.
}
\description{
Interne Hilfsfunktion von analysiereFrimen()
}
\details{
Berechnet die abnormale Returns pro Firma und Tag.
}
\author{
Harald Fiedler
}

