% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_varHat_epsilon_it.R
\name{calc_varHat_epsilon_it}
\alias{calc_varHat_epsilon_it}
\title{calc_varHat_epsilon_it}
\usage{
calc_varHat_epsilon_it(Firma, estimationWindow, eventWindow, Market)
}
\arguments{
\item{Wird}{aus dem Wrapper durchgereicht}
}
\value{
So wie ich das lese ist das wohl ein array von numerics
}
\description{
interne Funktion von analysiereFirmen()
}
\details{
Der Name ist selbstredend
}
\author{
Harald Fiedler
}

