% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventPlot.R
\name{EventPlot}
\alias{EventPlot}
\title{EventPlot}
\usage{
EventPlot(abnormals, KIleft, KIright, left = left)
}
\arguments{
\item{abnormals}{(vermutlich) ein numeric array mit abnormal returns, die gezeichnet werden sollen.}

\item{KIleft}{die zugehörigen unteren KI-Schranken}

\item{KIright}{die zugehörigen oberen kritischen Werte}

\item{left}{gibt an, wie viele Tage vor dem kritischen Event das Fenster anfangen soll. Dieser Wert wird tatsöchlich nur genutzt, um eine Vertikale zu zeichnen, die den DDay markiert.}
}
\description{
In Anlehnung an Fried ein Konfidenzband und dazu die CARs.
}
\author{
Harald Fiedler
}

