% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusioniereFirmenCARsMitMiSu03Daten.R
\name{fusioniereFirmenCARsMitMiSu03Daten}
\alias{fusioniereFirmenCARsMitMiSu03Daten}
\title{fusioniereFirmenCARsMitMiSu03Daten}
\usage{
fusioniereFirmenCARsMitMiSu03Daten(Firmenanalyse, MiSu03Daten)
}
\arguments{
\item{Firmenanalyse}{list die aus analysiereFirmen() gebildet wird}

\item{SU03Daten}{data.frame das man mit lese2016Q1MiSu03Daten(Event=eventdate) erhält}
}
\description{
fügt den SU03Daten die CAR hinzu
}
\details{
Die CAR_it werden an die SU03Daten gemerged mit t=letzter Tag des betrachteten EventWindows
}
\author{
Harald Fielder
}

