% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lese2016Q1MiSuDaten.R
\name{lese2016Q1MiSu03Daten}
\alias{lese2016Q1MiSu03Daten}
\title{lese2016Q1MiSu03Daten}
\usage{
lese2016Q1MiSu03Daten(Event)
}
\arguments{
\item{Event}{numeric der Länge 1 aus c(260, 555, 646, 809, 958)}
}
\value{
data.frame
}
\description{
Lädt die Daten ein
}
\details{
Lädt die Daten ein als interne Daten
}
\author{
Harald Fiedler
}

