% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadSulzbachDaten.R
\name{loadSulzbachDaten}
\alias{loadSulzbachDaten}
\title{loadSulbachDaten}
\usage{
loadSulzbachDaten(Gruppe, inProzent)
}
\arguments{
\item{Gruppe}{String der L채nge 1. Entweder "A", "B", "C", "AandB", oder "AxorB". Es handelt sich dabei um Firmengruppen, die unterschiedlich mit Pensionsrisiken umgehen.}

\item{inProzent}{Boolean der L채nge 1. Gibt an, ob die Analyse auf Basis absoluter Aktienkurse stattfindet, oder in relaitiven Prozent채nderungen zum ersten Tag des Event-Windows}
}
\description{
Stellt Datens채tze bereit
}
\details{


Stellt unter ISIN die zu untersuchenden Firmenindex bereit und unter KURS alle Aktienkurse
}
\author{
Harald Fiedler
}

