% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MacKinlayTable.R
\name{MacKinlayTable}
\alias{MacKinlayTable}
\title{MacKinlayTable}
\usage{
MacKinlayTable(Marktanalyse, LaTeX = TRUE)
}
\arguments{
\item{Marktanalyse}{ist ein Objekt, das von agggregiereFirmen() erzeugt wurde}

\item{LaTeX}{boolescher Entschieder, ob xtable genutzt werden soll, oder das Ergebnis als data.frame gegeben werden soll}
}
\value{
xtable-Objekt oder data.frame mit zwei Spalten und Fancy rownames
}
\description{
Erstellt Tabelle mit AR und CAR auf aggrtegierter Ebene
}
\author{
Harald Fiedler
}

