% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFirma.R
\name{plotFirma}
\alias{plotFirma}
\title{plotFirma}
\usage{
plotFirma(Firmenanalyse, FirmaISIN = "FI0009010391", DDay = 5)
}
\description{
Wrapper für EventPlot.
}
\details{
Plottet einzelne Firmen in ihren individuellen Konfidenzbänder
}
\author{
Harald FIedler
}

