% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMarkt.R
\name{plotMarkt}
\alias{plotMarkt}
\title{plotMarkt}
\usage{
plotMarkt(Marktanalyse, DDay = 5)
}
\description{
Wrapper für EventPlot.
}
\details{
Plottet den Markt in seinem  Konfidenzband
}
\author{
Harald FIedler
}

