% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFirma.R
\name{plotFirma}
\alias{plotFirma}
\title{plotFirma}
\usage{
plotFirma(Firmenanalyse, FirmaISIN = "FI0009010391", DDay)
}
\description{
Wrapper für EventPlot.
}
\details{
Plottet einzelne Firmen in ihren individuellen Konfidenzbänder
}
\author{
Harald FIedler
}

